+++ books/models/book.py
@@ -26,7 +26,7 @@ class Book(models.Model):
     image = models.ImageField(upload_to="book_pics", default="default.png", help_text="Min. size 300x300")
     publication_year = models.IntegerField()
     description = models.TextField(null=True, blank=True)
-    rating = models.IntegerField()
+    rating = models.DecimalField(max_digits=3, decimal_places=1)
     genre = models.ForeignKey(Genre, on_delete=models.DO_NOTHING, related_name='books')
     authors = models.ManyToManyField(Author, related_name='books')
